server: you cannot delete dish if it is used in menu

+++ server/src/main/java/com/ifmo/ratatoile/service/DishService.kt
@@ -5,15 +5,18 @@ import com.ifmo.ratatoile.dao.toDto
 import com.ifmo.ratatoile.dto.DishCreateRequestDto
 import com.ifmo.ratatoile.dto.DishDto
 import com.ifmo.ratatoile.dto.DishesDto
+import com.ifmo.ratatoile.exception.BadRequestException
 import com.ifmo.ratatoile.exception.NotFoundException
 import com.ifmo.ratatoile.repository.DishRepository
+import com.ifmo.ratatoile.repository.MenuEntryRepository
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import java.math.BigDecimal
 
 @Service
 class DishService(
-  private val dishRepository: DishRepository
+  private val dishRepository: DishRepository,
+  private val menuEntryRepository: MenuEntryRepository
 ) {
 
   fun getDishAsEntity(id: Int): Dish {
@@ -41,7 +44,16 @@ class DishService(
 
   fun deleteDish(id: Int): DishDto {
     val i = getDishAsEntity(id)
+    if (isDishUsedInMenu(id)) {
+      throw BadRequestException("Dish id=$id is used in menu")
+    }
     dishRepository.deleteById(id)
     return i.toDto()
   }
+
+  private fun isDishUsedInMenu(dishId: Int): Boolean {
+    val dishIDs = menuEntryRepository.findAll().map { it.dishId }
+    return dishIDs.contains(dishId)
+  }
+
 }
